pyproject.toml
[tool.poetry]
name = "django_forms_plus"
version = "0.1.0"
description = "React-powered forms for Django"
license = "MIT"
authors = ["Vladimir Sklyar <versus.post@gmail.com>"]
repository = "https://github.com/versusbassz/django_forms_plus"
readme = "README.md"
keywords = ["django", "forms", "react"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: JavaScript",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: User Interfaces",
]
packages = [
    { include = "django_forms_plus" }
]
include = ["django_forms_plus/static/**/*"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
Django = "^3.2"
django-extensions = "^3.2.1"
mkdocs = "^1.4.2"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
